Implement player list (#11)
* organizando pacotes

* implement player list

---------

+++ lib/model/player.dart
@@ -4,7 +4,7 @@ class Player {
   String name;
   Position principalPosition;
   Position? secondaryPosition;
-  int overall;
+  double overall;
 
   Player({
     required this.name,
